+++ src/twitch/TwitchDashboard.cpp
@@ -66,8 +66,9 @@ bool TwitchDashboard::setup() {
     // Create welcome label
     std::string welcomeText = "Welcome, " + channelName + "!";
     m_welcomeLabel = CCLabelBMFont::create(welcomeText.c_str(), "bigFont.fnt");
-    m_welcomeLabel->setPosition(layerSize.width / 5, layerSize.height - 22);
+    m_welcomeLabel->setPosition(layerSize.width / 5 - 5.f, layerSize.height - 22.f);
     m_welcomeLabel->setScale(0.3f);
+    m_welcomeLabel->setAlignment(kCCTextAlignmentLeft);
     m_welcomeLabel->setID("welcome-label");
     m_mainLayer->addChild(m_welcomeLabel);