+++ changelog.md
@@ -1,5 +1,5 @@
 # v0.1.5-beta.1
-- Added **Sound Effect** Event 
+- Added **Sound Effect** & **Stop All Sounds** Events
 - Changed the **Wait** Action value to use float value instead of just integer.
 # v0.1.4-beta.1
 - Added **Search Event** and sorted the events in an alphabetical order